Overall Ratings Function
Function that collects the total meta_values for a given meta_key for all comments as reviews for an individual post (while on that post page). 

This function can be used within the loop by being called as overall_ratings(); and its output is a number. If the number is 0, the function will display "Unrated" as the result.